BAAS-19768: update goja to latest March 2023 (part 2) (#79)
* Fixed panic in newArrayFromIter when the iterator is already closed. Fixes #375
* Fixed panic when parsing invalid object property keys. Fixes #376.
* Fixed accidental shadowing in the else branches of type assertion
* Fixed defineProperty("length") for arrays. Improved detection of non-standard array configurations. Upgraded tc39 tests.
* Return true values of struct fields or reflect slice elements, rather than pointers to them. Closes #378.
* Upgraded dependencies. Closes #380.
* Implemented exponentiation expressions. Closes #381.
* Enabled tests that use ** operator. Some array fixes as a result.
* Implemented nullish coalescing operator (??). Closes #382.
* Implemented `{Array,String,%TypedArray%}.prototype.at` (#384)e7c2872c8)
* Fixed callee expressions in optional chains. Fixes #385.
* Do not use fmt.Sprintf() for plain error strings. Fixes #388.
* Implemented 'copy-on-change' mechanism for inner compound values. Fixes #403.
* Fixed objectGoReflect equality. See #403
* Don't clear interrupt until the stack is empty (#405)
* test: skip Promise based tests

+++ object_dynamic.go
@@ -132,15 +132,15 @@ func (r *Runtime) NewDynamicArray(a DynamicArray) *Object {
 	return v
 }
 
-func (*dynamicObject) sortLen() int64 {
+func (*dynamicObject) sortLen() int {
 	return 0
 }
 
-func (*dynamicObject) sortGet(i int64) Value {
+func (*dynamicObject) sortGet(i int) Value {
 	return nil
 }
 
-func (*dynamicObject) swap(i int64, i2 int64) {
+func (*dynamicObject) swap(i int, i2 int) {
 }
 
 func (*dynamicObject) className() string {
@@ -530,15 +530,15 @@ func (b *baseDynamicObject) MemUsage(ctx *MemUsageContext) (uint64, error) {
 	return b.val.MemUsage(ctx)
 }
 
-func (a *dynamicArray) sortLen() int64 {
-	return int64(a.a.Len())
+func (a *dynamicArray) sortLen() int {
+	return a.a.Len()
 }
 
-func (a *dynamicArray) sortGet(i int64) Value {
-	return a.a.Get(int(i))
+func (a *dynamicArray) sortGet(i int) Value {
+	return a.a.Get(i)
 }
 
-func (a *dynamicArray) swap(i int64, j int64) {
+func (a *dynamicArray) swap(i int, j int) {
 	x := a.sortGet(i)
 	y := a.sortGet(j)
 	a.a.Set(int(i), y)
